organize api folder more clearly
