+++ languages/node/api/commerce-data/index.js
@@ -1,0 +1,9 @@
+module.exports = {
+  ...require('./orders'),
+  ...require('./products'),
+  ...require('./customers'),
+  ...require('./locations'),
+  ...require('./inventories'),
+  ...require('./collections'),
+  ...require('./bulkOperations'),
+};
