+++ languages/node/api/index.js
@@ -1,13 +1,4 @@
 module.exports = {
-  returnActions: require('./returnActions'),
-  returnData: require('./returnData'),
-  webhooks: require('./webhooks'),
-  itemGrading: require('./itemGrading'),
-  cart: require('./cart'),
-  destinations: require('./destinations'),
-  blocklist: require('./blocklist'),
-  allowlist: require('./allowlist'),
-  fraudReports: require('./fraudReports'),
-  happyReturns: require('./happyReturns'),
-  labelRequests: require('./labelRequests'),
+  ...require('./platform'),
+  ...require('./commerce-data'),
 };
