+++ languages/node/api/platform/destinations.js
@@ -1,4 +1,4 @@
-const { loopApiRequest } = require('./core');
+const { loopApiRequest } = require('../core');
 
 /**
  * Fetch all return destinations for the shop
