+++ languages/node/api/platform/happyReturns.js
@@ -1,4 +1,4 @@
-const { loopApiRequest } = require('./core');
+const { loopApiRequest } = require('../core');
 
 /**
  * Get high-level info about a specific Happy Returns shipment.
