+++ languages/node/api/platform/index.js
@@ -1,0 +1,12 @@
+module.exports = {
+  ...require('./returnData'),
+  ...require('./returnActions'),
+  ...require('./webhooks'),
+  ...require('./destinations'),
+  ...require('./labelRequests'),
+  ...require('./listings'),
+  ...require('./fraudReports'),
+  ...require('./happyReturns'),
+  ...require('./itemGrading'),
+  ...require('./cart'),
+};
