+++ languages/node/api/platform/itemGrading.js
@@ -1,4 +1,4 @@
-const { loopApiRequest } = require('./core');
+const { loopApiRequest } = require('../core');
 
 /**
  * Grade the condition of return line items (up to 20 at a time).
