+++ languages/node/api/platform/listings.js
@@ -1,4 +1,4 @@
-const { loopApiRequest } = require('./core');
+const { loopApiRequest } = require('../core');
 
 //
 // 🔒 Blocklist
