+++ languages/node/api/platform/returnData.js
@@ -1,4 +1,4 @@
-const { loopApiRequest } = require('./core');
+const { loopApiRequest } = require('../core');
 
 /**
  * Get return details using return_id, order_id, or order_name
