+++ languages/node/api/platform/webhooks.js
@@ -1,4 +1,4 @@
-const { loopApiRequest } = require('./core');
+const { loopApiRequest } = require('../core');
 
 /**
  * Retrieve all webhook subscriptions (programmatic + admin-created)
